implementacion test parqueadero

+++ domain/src/main/java/com/example/domain/valueobject/CobroServicio.kt
@@ -2,12 +2,11 @@ package com.example.domain.valueobject
 
 import com.example.domain.entity.Carro
 import com.example.domain.entity.Moto
-import kotlin.math.ceil
 
 interface CobroServicio {
 
-    fun cobroTarifaMoto(duracionServicio: Double, moto: Moto): Int
+    fun cobroTarifaMoto(duracionServicio: Int, moto: Moto): Int
 
-    fun cobroTarifaCarro(duracionServicio: Double, carro: Carro): Int
+    fun cobroTarifaCarro(duracionServicio: Int, carro: Carro): Int
 
 }
